# Change ph5tostationxml.py to evaluate each experiment in a separate thread