I modified ph5tostationxml.py to evaluate each experiment in a separate thread. In some initial testing this offered a significant performance improvement for request with several experiments.

Tests using CURL:
Single Threaded:
`http://service.iris.edu/ph5beta/station/1/query?network=4C,ZI,8A&format=text `
took 18.847675 to 20.693347 seconds

Multithreaded:
`http://service.iris.edu/ph5beta/station/1/query?network=4C,ZI,8A&format=text `
took 11.394308 to 12.170499 seconds

@derick-hess  Please note that this pull request is based on the ph5utils branch and should be merged before ph5utils is merged to master.