@nick-iris 2017-05-19
It should be noted that when I say "multi-threaded" I actually mean "multiple-processes". The `pool.map` function is used to read each experiment in a separate system process.