@nick-iris 2017-05-19
Yeah before this modification a request for all networks in the station service would time out. I'm also planning to do something similar with ph5toexml.py.